+++ example/native/App.tsx
@@ -1,0 +1,34 @@
+import { StatusBar } from "expo-status-bar";
+import { Alert, StyleSheet, View } from "react-native";
+import { TreegeRenderer } from "treege/renderer-native";
+import flows from "~/example/json/treege.json";
+import { Flow } from "@/shared/types/node";
+
+export default function App() {
+  const handleSubmit = (values: any) => {
+    Alert.alert("Form Submitted", JSON.stringify(values, null, 2), [{ text: "OK" }]);
+  };
+
+  return (
+    <View style={styles.container}>
+      <TreegeRenderer
+        flows={flows  as Flow}
+        onSubmit={handleSubmit}
+        contentContainerStyle={styles.scrollContent}
+      />
+      <StatusBar style="auto" />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: "#F9FAFB",
+    flex: 1,
+  },
+  scrollContent: {
+    flex: 1,
+    justifyContent: "center",
+    paddingHorizontal: 16,
+  },
+});